requirements copy.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
shapely>=2.0.0
tqdm>=4.65.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0.0
opencv-python>=4.7.0
requests>=2.31.0
